+++ nova/src/Core/Pages/Http/Controllers/PageController.php
@@ -26,83 +26,115 @@ class PageController extends BaseController {
 
 	public function index()
 	{
-		$this->view = 'admin/pages/pages';
-		$this->jsView = 'admin/pages/pages-js';
+		if ($this->user->can(['page.create', 'page.edit', 'page.remove']))
+		{
+			$this->view = 'admin/pages/pages';
+			$this->jsView = 'admin/pages/pages-js';
+		}
+
+		return $this->errorUnauthorized("You do not have permission to manage pages.");
 	}
 
 	public function create()
 	{
-		$this->view = 'admin/pages/page-create';
-		$this->jsView = 'admin/pages/page-create-js';
+		if ($this->user->can('page.create'))
+		{
+			$this->view = 'admin/pages/page-create';
+			$this->jsView = 'admin/pages/page-create-js';
 
-		$this->data->httpVerbs = [
-			'GET' => 'GET',
-			'POST' => 'POST',
-			'PUT' => 'PUT',
-			'DELETE' => 'DELETE',
-		];
+			$this->data->httpVerbs = [
+				'GET' => 'GET',
+				'POST' => 'POST',
+				'PUT' => 'PUT',
+				'DELETE' => 'DELETE',
+			];
 
-		$this->data->menus[0] = "No menu";
-		$this->data->menus += $this->menuRepo->listAll('name', 'id');
+			$this->data->menus[0] = "No menu";
+			$this->data->menus += $this->menuRepo->listAll('name', 'id');
+		}
+
+		return $this->errorUnauthorized("You do not have permission to create pages.");
 	}
 
 	public function store(CreatePageRequest $request)
 	{
-		// Create the page
-		$page = $this->repo->create($request->all());
+		if ($this->user->can('page.create'))
+		{
+			// Create the page
+			$page = $this->repo->create($request->all());
 
-		// Fire the event
-		event(new Events\PageWasCreated($page));
+			// Fire the event
+			event(new Events\PageWasCreated($page));
 
-		// Set the flash message
-		flash()->success("Page has been created. Don't forget to update your menu(s) with your new page!");
+			// Set the flash message
+			flash()->success("Page has been created. Don't forget to update your menu(s) with your new page!");
 
-		return redirect()->route('admin.pages');
+			return redirect()->route('admin.pages');
+		}
+
+		return $this->errorUnauthorized("You do not have permission to create pages.");
 	}
 
 	public function edit($pageId)
 	{
-		$this->view = 'admin/pages/page-edit';
-		$this->jsView = 'admin/pages/page-edit-js';
+		if ($this->user->can('page.edit'))
+		{
+			$this->view = 'admin/pages/page-edit';
+			$this->jsView = 'admin/pages/page-edit-js';
 
-		$this->data->page = $this->repo->find($pageId);
-		
-		$this->data->httpVerbs = [
-			'GET' => 'GET',
-			'POST' => 'POST',
-			'PUT' => 'PUT',
-			'DELETE' => 'DELETE',
-		];
+			$this->data->page = $this->repo->find($pageId);
+			
+			$this->data->httpVerbs = [
+				'GET' => 'GET',
+				'POST' => 'POST',
+				'PUT' => 'PUT',
+				'DELETE' => 'DELETE',
+			];
 
-		$this->data->menus[0] = "No menu";
-		$this->data->menus += $this->menuRepo->listAll('name', 'id');
+			$this->data->menus[0] = "No menu";
+			$this->data->menus += $this->menuRepo->listAll('name', 'id');
+		}
+
+		return $this->errorUnauthorized("You do not have permission to edit pages.");
 	}
 
 	public function update(EditPageRequest $request, $pageId)
 	{
-		// Update the page
-		$page = $this->repo->update($pageId, $request->all());
+		if ($this->user->can('page.edit'))
+		{
+			// Update the page
+			$page = $this->repo->update($pageId, $request->all());
 
-		// Fire the event
-		event(new Events\PageWasUpdated($page));
+			// Fire the event
+			event(new Events\PageWasUpdated($page));
 
-		// Set the flash message
-		flash()->success("Page has been updated.");
+			// Set the flash message
+			flash()->success("Page has been updated.");
 
-		return redirect()->route('admin.pages');
+			return redirect()->route('admin.pages');
+		}
+
+		return $this->errorUnauthorized("You do not have permission to edit pages.");
 	}
 
 	public function remove($pageId)
 	{
 		$this->isAjax = true;
 
-		// Grab the page we're removing
-		$page = $this->repo->find($pageId);
+		if ($this->user->can('page.remove'))
+		{
+			// Grab the page we're removing
+			$page = $this->repo->find($pageId);
 
-		// Build the body based on whether we found the page or not
-		$body = ($page)
-			? view(locate('page', 'admin/pages/page-remove'), compact('page'))
-			: alert('danger', "Page not found.");
+			// Build the body based on whether we found the page or not
+			$body = ($page)
+				? view(locate('page', 'admin/pages/page-remove'), compact('page'))
+				: alert('danger', "Page not found.");
+		}
+		else
+		{
+			$body = alert('danger', "You do not have permission to remove pages.");
+		}
 
 		return partial('modal-content', [
 			'header' => "Remove Page",
@@ -113,16 +145,21 @@ class PageController extends BaseController {
 
 	public function destroy(RemovePageRequest $request, $pageId)
 	{
-		// Delete the page
-		$page = $this->repo->delete($pageId);
+		if ($this->user->can('page.remove'))
+		{
+			// Delete the page
+			$page = $this->repo->delete($pageId);
 
-		// Fire the event
-		event(new Events\PageWasDeleted($page->name, $page->key, $page->uri));
+			// Fire the event
+			event(new Events\PageWasDeleted($page->name, $page->key, $page->uri));
 
-		// Set the flash message
-		flash()->success("Page has been removed.");
+			// Set the flash message
+			flash()->success("Page has been removed.");
 
-		return redirect()->route('admin.pages');
+			return redirect()->route('admin.pages');
+		}
+
+		return $this->errorUnauthorized("You do not have permission to remove pages.");
 	}
 
 	public function checkPageKey()
